refactor: separate service at `lib/product-service`
