+++ lib/product-service/constant.ts
@@ -1,0 +1,1 @@
+export const LAMBDA_FOLDER_PATH = "./lambda";
