+++ lib/product-service/lambda/get-products-list/handler.ts
@@ -1,0 +1,5 @@
+import { getProducts } from "../mock";
+
+export async function getProductsList() {
+  return getProducts();
+}
