+++ lib/product-service/lambda/index.ts
@@ -1,0 +1,2 @@
+export { getProductsById } from "./get-products-by-id/handler";
+export { getProductsList } from "./get-products-list/handler";
