+++ lib/product-service/lambda/mock.ts
