+++ lib/product-service/lambda/types.ts
